Breaking Change: Remove dataset handling from zookeeper
This removes most of the environment specific dataset handling from `zookeeper`.
The obvious downside is that this requires more code in userspace and users need to optimize the data pipeline themselves.

The upside of this approach is that it adds a lot more flexibility to `zookeeper` without us needing to support every `tf.data` feature through `kwargs`. E.g. we've seen OOM errors due to `AUTOTUNE` in `.prefetch()` in a not carefully configured Kubernetes environment. All of the functionality that was previously included in `zookeeper` can be easily added with just a few lines of code, so I think this will make `zookeeper` more useful in different computing environments.

In the future, I'd like to remove the current dataset caching functionality as well and replace it with the in memory caching provided by TensorFlow datasets in in the `tfds-nightly` release. The current `tf.data` disk caching functionality is quite brittle (https://github.com/tensorflow/tensorflow/issues/28860 https://github.com/tensorflow/tensorflow/issues/28798) so it is not really useful in this context anyway.