Perform component configuration in a breadth-first order.
Previously, components were greedily configured in a depth-first
traversal of the component tree, meaning that instances which were
inherited could get configured at the 'wrong' (non-topmost) level.
This was a source of hard-to-identify bugs - e.g. closes #226.

Unfortunately, breadth-first configuration is incompatible with the
user-configurable `__component_configure__` method that overrides
configuration (because the order of recursive configuration now
needs to be externally imposed), so `__component_configure__` is
replaced with `__pre_configure__` (name analagous to the existing
`__post_configure__`), a method which takes a single argument - a
configuration dict - and can return a *new* configuration dict that
will be later used for configuring the instance. This method affords
the same abilities/powers/flexibility of the
`__component_configure__` method it replace.